+++ src/tests/guardian_wrapper_test.rs
@@ -76,6 +76,7 @@ fn create_test_options() -> CreateDBOptions {
         close_func: None,
         store_specific_opts: None,
         doc_ticket: None,
+        read_only: None,
     }
 }